refactor: move shield module out of auth module
