+++ src/shield/rules/post.rules.ts
@@ -1,7 +1,7 @@
 import { rule } from "graphql-shield";
-import { UNAUTHORIZED } from "../../../common/common.constants";
-import { Context } from "../../../context/context.service";
-import { GroupPrivacy } from "../../../groups/group-configs/models/group-config.model";
+import { UNAUTHORIZED } from "../../common/common.constants";
+import { Context } from "../../context/context.service";
+import { GroupPrivacy } from "../../groups/group-configs/models/group-config.model";
 
 export const isOwnPost = rule()(
   async (_parent, args, { user, services: { usersService } }: Context) => {
