+++ src/shield/shield.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { EventsModule } from "../../events/events.module";
+import { EventsModule } from "../events/events.module";
 import { ShieldService } from "./shield.service";
 
 @Module({
