+++ src/shield/shield.permissions.ts
@@ -1,5 +1,5 @@
 import { allow, and, not, or, shield } from "graphql-shield";
-import { FORBIDDEN } from "../../common/common.constants";
+import { FORBIDDEN } from "../common/common.constants";
 import { hasValidRefreshToken, isAuthenticated } from "./rules/auth.rules";
 import {
   canApproveGroupMemberRequests,
